refactor: project dir
