+++ src/App.tsx
@@ -1,12 +1,12 @@
-import AppRoutes from './routes';
+import AppRoutes from './components/Navbar';
 
 function App() {
-  return (
-    <>
-      {/* Modals Here */}
-      <AppRoutes />
-    </>
-  );
+	return (
+		<>
+			{/* Modals Here */}
+			<AppRoutes />
+		</>
+	);
 }
 
 export default App;
