--- src/components/BackIcon.tsx
@@ -1,25 +1,0 @@
-interface BackIconProps {
-  className?: string;
-}
-
-const BackIcon: React.FC<BackIconProps> = ({ className }) => {
-  return (
-    <svg
-      width="40"
-      height="40"
-      viewBox="0 0 40 40"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      className={className}
-    >
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M7.175 19.375L12.1 24.3C12.2842 24.4716 12.432 24.6786 12.5345 24.9086C12.6369 25.1386 12.692 25.3869 12.6965 25.6387C12.7009 25.8904 12.6546 26.1405 12.5603 26.374C12.466 26.6074 12.3256 26.8195 12.1476 26.9976C11.9695 27.1756 11.7575 27.316 11.524 27.4103C11.2905 27.5046 11.0405 27.5509 10.7887 27.5465C10.5369 27.542 10.2887 27.4869 10.0587 27.3844C9.82866 27.282 9.62166 27.1342 9.45 26.95L1.325 18.825L0 17.5L1.325 16.175L9.45 8.04998C9.80544 7.71878 10.2756 7.53847 10.7613 7.54705C11.2471 7.55562 11.7105 7.7524 12.0541 8.09593C12.3976 8.43946 12.5944 8.90292 12.6029 9.38868C12.6115 9.87443 12.4312 10.3445 12.1 10.7L7.175 15.625H31.875C34.0299 15.625 36.0965 16.481 37.6202 18.0047C39.144 19.5285 40 21.5951 40 23.75C40 25.9049 39.144 27.9715 37.6202 29.4952C36.0965 31.019 34.0299 31.875 31.875 31.875H26.875C26.3777 31.875 25.9008 31.6774 25.5492 31.3258C25.1975 30.9742 25 30.4973 25 30C25 29.5027 25.1975 29.0258 25.5492 28.6742C25.9008 28.3225 26.3777 28.125 26.875 28.125H31.875C33.0353 28.125 34.1481 27.664 34.9686 26.8436C35.7891 26.0231 36.25 24.9103 36.25 23.75C36.25 22.5897 35.7891 21.4769 34.9686 20.6564C34.1481 19.8359 33.0353 19.375 31.875 19.375H7.175Z"
-        fill="#1E1E1E"
-      />
-    </svg>
-  );
-};
-
-export default BackIcon;
